drivers-dma-remove-unused-support-for-memset-operations-v3
This is an old patch that I ported to current kernels.  I fixed rejects
and verified that it compiles for cases that hit rejects but unfortunately
I forgot to check for new code related to DMA_MEMSET functionality.
Fortunately it seems that the problematic places are in
plat-orion/common.c and mv_xor.c.  They are fixed in the attached patch.

Signed-off-by: Bartlomiej Zolnierkiewicz <b.zolnierkie@samsung.com>
Cc: Kyungmin Park <kyungmin.park@samsung.com>
Cc: Vinod Koul <vinod.koul@intel.com>
Cc: Dan Williams <djbw@fb.com>
Cc: Tomasz Figa <t.figa@samsung.com>
Cc: Herbert Xu <herbert@gondor.apana.org.au>
Cc: Olof Johansson <olof@lixom.net>
Cc: Kevin Hilman <khilman@linaro.org>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>