memcg: fix build error if CONFIG_MEMCG_KMEM=n
Fix this build error:

mm/built-in.o: In function `mem_cgroup_css_free':
memcontrol.c:(.text+0x5caa6): undefined reference to 'mem_cgroup_sockets_destroy'

Maybe it's better to add memcg_destroy_kmem(), to pair with memcg_init_kmem().

This patch can be folded into "memcg: use css_get/put when charging/uncharging kmem".

Rename the patch as it is a follow-up for the mmotm tree:
"memcg-use-css_get-put-when-charging-uncharging-kmem-fix-2.patch"

See [1] for the original patch and follow the discussion in [2].

[1] http://marc.info/?l=linux-mm&m=137285590123430&w=2
[2] http://marc.info/?t=137285378500005&r=1&w=2

Reported-by: Fengguang Wu <fengguang.wu@intel.com>
Reported-by: Stephen Rothwell <sfr@canb.auug.org.au>
Reported-by: Sedat Dilek <sedat.dilek@gmail.com>
Tested-by: Sedat Dilek <sedat.dilek@gmail.com>
Acked-by: Michal Hocko <mhocko@suse.cz>
Signed-off-by: Li Zefan <lizefan@huawei.com>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>