Link: Prevent loss of backlight level during disable/enable sequence.
Tracking of whether user space had explicitly requested a non-zero
backlight level was broken, such that the backlight level was set to
0 when disabling the backlight.  Re-enabling would then still produce
a black screen.

Prior to CL 61336 this was hidden by the kernel driver resetting the
level to 255 when enabling the backlight.  While wrong, this value did
not produce a black screen.

The Link backlight driver should now behave like the non-Link one, and
retain the backlight level across a disable/enable sequence.

BUG=chromium:259457
TEST=No blank screen on log out. No bright flash during suspend.

Change-Id: Iea708db7c8f3ff896ec17ca2924419872ff61199
Reviewed-on: https://gerrit.chromium.org/gerrit/61782
Tested-by: Stuart Abercrombie <sabercrombie@chromium.org>
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>
Commit-Queue: Stuart Abercrombie <sabercrombie@chromium.org>