ACPI: remove panic in case hardware signature has changed after S4
Some BIOSes change hardware signatures based on the state of a
laptop's lid.  When the lid is closed, the touchpad is disabled and
the checksum changes.

Windows 8 no longer aborts resume if the checksum has changed, so we
need to follow it in that respect, because many new machines are only
tested with Windows 8.

[rjw: Changelog]
Signed-off-by: Oliver Neukum <oneukum@suse.de>
No-objections-from: Thomas Renninger <trenn@suse.de>
Signed-off-by: Rafael J. Wysocki <rafael.j.wysocki@intel.com>