drm/i915: Replace racy object bookkeeping
Now that we track objects for their entire lifetime in a list, we can
move the cost of the bookkeeping to the infrequent query of
i915_gem_objects. This also removes the race where we would increment the
global object count and size without holding any locks.

Bugzilla: https://bugs.freedesktop.org/show_bug.cgi?id=67121
Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>
Signed-off-by: Daniel Vetter <daniel.vetter@ffwll.ch>