UPSTREAM: video/hmdi: Clear the whole incoming buffer, not just the infoframe size
If the user if this API is providing a bigger buffer than the infoframe
size, it could be for a could reason. For instance it could be because
it gives the buffer that will be written to the hardware, up to the
maximum of an infoframe size.

Instead of just zeroing up to the infoframe size, let's zero the whole
incoming buffer as those extra bytes are also used to compute the
ECC and need to be 0.

Signed-off-by: Damien Lespiau <damien.lespiau@intel.com>
Acked-by: Dave Airlie <airlied@gmail.com>
Reviewed-by: Alex Deucher <alexander.deucher@amd.com>
Reviewed-by: Ville Syrjälä <ville.syrjala@linux.intel.com>
Signed-off-by: Daniel Vetter <daniel.vetter@ffwll.ch>
(cherry picked from commit 3b390f62674eb70522057e58b1787f2cfa2f57c1)

Signed-off-by: Wayne Boyer <wayne.boyer@intel.com>

BUG=chrome-os-partner:26772
TEST=Connect Rambi via HDMI to ASUS VE228 monitor and verify correct
output on the monitor.

Change-Id: I84bceea4ff404bb1e15c05b6b146a4a81ef8d925
Reviewed-on: https://chromium-review.googlesource.com/189622
Reviewed-by: Jim Bride <jim.bride@intel.com>
Reviewed-by: Benson Leung <bleung@chromium.org>
Commit-Queue: Wayne Boyer <wayne.boyer@intel.com>
Tested-by: Wayne Boyer <wayne.boyer@intel.com>