CHROMIUM: exynos: hdmi: Dereference ddc_node on use.
Order of initialization of exynos_hdmi and ddc is not currently guaranteed.
This patch allows the DDC channel to be used even if it happens to be
initialized after hdmi.

This patch should be dropped when HDMI driver initialization is guaranteed to
happen after DDC channel on which it depends.

BUG=chrome-os-partner:18395
TEST=Use external HDMI port on spring and observe that it still works.

Original-Change-Id: Iecbe80fa1a37fd83b2c31eac3c0b7e5284b8c033
Signed-off-by: Jeremy Thorpe <jeremyt@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/65014
(cherry picked from commit 3afbca24a5037dca2349d1ae2b9c89675e546287)

Change-Id: I161b7ef1817cd8af588d0cfdcd9ea7867c86162e
Reviewed-on: https://gerrit.chromium.org/gerrit/66813
Reviewed-by: Jeremy Thorpe <jeremyt@google.com>
Tested-by: Jeremy Thorpe <jeremyt@google.com>