drm: exynos : restore lost configuration on dpms off
In the case of restart ui or user signng out, all
the configuration is lost in DPMS OFF as a result
for proper functioning of display, the lost
configuration needs to be restored, this patch
restores it.

This patch also  modifies fimd, as in the case
if there are no page flips, the same may be required.

BUG=chrome-os-partner:21758
TEST=suspend_stress_test,hotplug restart ui
     and user sign out with HDMI connected on peach

Change-Id: If09e158016c67212f7505a136c4d288884b141de
Signed-off-by: Shirish S <s.shirish@samsung.com>
Previous-Reviewed-on: https://gerrit.chromium.org/gerrit/65181
(cherry picked from commit 1591e40e5ee0fa9afcfc27457c85b97fa7cb7ef0)
Reviewed-on: https://chromium-review.googlesource.com/167300
Reviewed-by: Sean Paul <seanpaul@google.com>
Reviewed-by: Douglas Anderson <dianders@chromium.org>
Commit-Queue: Douglas Anderson <dianders@chromium.org>
Tested-by: Douglas Anderson <dianders@chromium.org>